Rename data_interface->usefull_functions
