+++ app/make_together.py
@@ -1,7 +1,7 @@
 import os
 import logging
 import shutil
-from dataset_interface import create_csv, write_in_csv
+from usefull_functions import create_csv, write_in_csv
 from get_way import get_path_normal, get_path_together
